ui: Refactor sidebar
Refactored sidebar by extracting a number of components into their own
classes, and also moved the timeline menu item out into the timeline
plugin.
