+++ ui/src/frontend/views/hiring_banner.ts
@@ -1,0 +1,44 @@
+// Copyright (C) 2026 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import m from 'mithril';
+import {AppImpl} from '../../core/app_impl';
+import {featureFlags} from '../../core/feature_flags';
+
+const HIRING_BANNER_FLAG = featureFlags.register({
+  id: 'showHiringBanner',
+  name: 'Show hiring banner',
+  description: 'Show the "We\'re hiring" banner link in the side bar.',
+  defaultValue: false,
+});
+
+export function shouldShowHiringBanner(app: AppImpl): boolean {
+  return app.isInternalUser && HIRING_BANNER_FLAG.get();
+}
+
+export const HiringBanner: m.Component = {
+  view() {
+    return m(
+      '.pf-hiring-banner',
+      m(
+        'a',
+        {
+          href: 'http://go/perfetto-open-roles',
+          target: '_blank',
+        },
+        "We're hiring!",
+      ),
+    );
+  },
+};
